+++ slacky/plugins/custom/__init__.py
@@ -2,3 +2,4 @@
 from .example import custom_example
 # Uncomment to load Deepfry Plugin
 # from .deepfry import deepfry
+from .stockpic import stockpic